pip-requirements.txt
django==1.5
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
redis==2.7.2